Added boilerplate for the backend.
To start development of the backend (which will be the preferred way
to use the software), the runserver.py file has been added to run a
(currently basic) API server using FastAPI (best served with uvicorn
in development). See also requirements.txt for environment generation
and README.md for more information.