Improved project structure removed FastAPI backend
The backend part using FastAPI has been removed as we are developing a
much more complex Django application (known as QuizApp) on which we plan
to incorporate the current project. As a consequence, we removed the
requirements.txt since the code works now using only standard library.

We have also changed some of the documentation pages (minor changes)
and changed the size of the page on the template.tex file.